+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes to the plugin.
 
 ### Fixed
 
+- Creating a variable that starts with `choice_` is now properly flagged as an error.
 - `*save_checkpoint` and `*restore_checkpoint` now don't give an error when they don't include a slot's name.
 - `*save_checkpoint` and `*restore_checkpoint` now autocomplete properly.
 - `choice_just_restored_checkpoint` is now properly treated as a built-in variable.